requirements.txt
pandas
scikit-learn
numpy
pyyaml
matplotlib
MetaTrader5
GitPython
